+++ app/models/incident_entity.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, String, UniqueConstraint, func
+from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, Integer, String, UniqueConstraint, func
+
 
 from app.db import Base
 
@@ -8,7 +9,8 @@ from app.db import Base
 class IncidentEntity(Base):
     __tablename__ = "incident_entities"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+
 
     incident_id = Column(BigInteger, ForeignKey("incidents.id", ondelete="CASCADE"), nullable=False)
     entity_id = Column(BigInteger, ForeignKey("entities.id", ondelete="CASCADE"), nullable=False)
